Extract plugin/git runtime-identity helpers into runtime_identity.py (WS5 seam)
Move the plugin- and git-identity provenance helpers out of engine.py into a
new runtime_identity.py module: _strip_metadata_scalar, _plugin_metadata (with
its _PLUGIN_METADATA manifest cache and _PLUGIN_ROOT), and _git_runtime_identity.
These feed the plugin_* fields of the runtime status/doctor payload. engine.py
imports the three symbols it references (_plugin_metadata, _git_runtime_identity,
_PLUGIN_ROOT); _strip_metadata_scalar and the cache stay internal to the module.
The move drops engine.py's now-unused `subprocess` import.

Tests that reached these through the engine module are retargeted to the new
module: the two _plugin_metadata tests patch/read runtime_identity._PLUGIN_METADATA
(the cache moved with its owner), and the two _git_runtime_identity packaging
tests patch runtime_identity.subprocess (engine no longer imports subprocess).

Behaviour-preserving refactor (WS5 engine decomposition). The moved functions are
byte-identical to their pre-move form; only their module-local logger name
changes. Rollback: git revert this commit restores the inline block; no data or
state migration involved.
