+++ engine.py
@@ -13,7 +13,6 @@ import logging
 import os
 import re
 import sqlite3
-import subprocess
 import threading
 import time
 import weakref
@@ -70,6 +69,11 @@ from .ingest_protection import (
     redact_sensitive_value,
     restore_ingest_payload_placeholders,
     sensitive_pattern_status,
+)
+from .runtime_identity import (
+    _PLUGIN_ROOT,
+    _git_runtime_identity,
+    _plugin_metadata,
 )
 from .schemas import (
     LCM_DESCRIBE,
@@ -183,8 +187,6 @@ def resolve_active_lcm_engine(session_id: str = "", conversation_id: str = "") -
                 _ACTIVE_ENGINES_BY_CONVERSATION_ID.pop(conversation_id, None)
     return None
 
-_PLUGIN_ROOT = Path(__file__).resolve().parent
-_PLUGIN_METADATA: dict[str, str] | None = None
 _SESSION_END_BUSY_TIMEOUT_MS = 50
 # ChatGPT Codex OAuth exposes provider-enforced context windows that can be
 # materially lower than the same model slug on direct OpenAI/OpenRouter routes.
@@ -254,78 +256,6 @@ def _is_codex_gpt55_route(model: str | None, provider: str | None) -> bool:
 _AUTO_FOCUS_MAX_TURNS = 3
 _AUTO_FOCUS_TURN_MAX_CHARS = 260
 _AUTO_FOCUS_MAX_CHARS = 700
-
-
-def _strip_metadata_scalar(value: str) -> str:
-    return value.strip().strip('"').strip("'")
-
-
-def _plugin_metadata() -> dict[str, str]:
-    """Return plugin identity from the loaded code tree.
-
-    Always re-read the manifest from disk when available so status tools reflect
-    hot-updated plugin checkouts even in long-lived Hermes processes.
-    """
-    global _PLUGIN_METADATA
-
-    metadata = {"name": "hermes-lcm", "version": "unknown"}
-    manifest = _PLUGIN_ROOT / "plugin.yaml"
-    try:
-        for line in manifest.read_text(encoding="utf-8").splitlines():
-            key, sep, raw_value = line.partition(":")
-            if not sep:
-                continue
-            key = key.strip()
-            if key in {"name", "version"}:
-                metadata[key] = _strip_metadata_scalar(raw_value)
-        _PLUGIN_METADATA = metadata
-        return dict(metadata)
-    except OSError:
-        logger.debug("LCM plugin manifest not readable at %s", manifest)
-
-    if _PLUGIN_METADATA is not None:
-        return dict(_PLUGIN_METADATA)
-    return dict(metadata)
-
-
-def _git_runtime_identity(root: Path) -> dict[str, Any]:
-    """Best-effort git identity for source checkouts.
-
-    Packaged installs may not have a `.git` directory. In that case the fields
-    stay empty instead of turning status/doctor into a git dependency.
-    """
-
-    if not (root / ".git").exists():
-        return {
-            "plugin_git_commit": "",
-            "plugin_git_branch": "",
-            "plugin_git_dirty": None,
-            "plugin_git_remote": "",
-        }
-
-    def _git(*args: str) -> str | None:
-        try:
-            result = subprocess.run(
-                ["git", "-C", str(root), *args],
-                check=False,
-                capture_output=True,
-                text=True,
-                timeout=1,
-            )
-        except (OSError, subprocess.SubprocessError) as exc:
-            logger.debug("LCM git identity probe failed at %s: %s", root, exc)
-            return None
-        if result.returncode != 0:
-            return None
-        return result.stdout.strip()
-
-    dirty_output = _git("status", "--porcelain")
-    return {
-        "plugin_git_commit": _git("rev-parse", "HEAD") or "",
-        "plugin_git_branch": _git("rev-parse", "--abbrev-ref", "HEAD") or "",
-        "plugin_git_dirty": None if dirty_output is None else bool(dirty_output),
-        "plugin_git_remote": _git("config", "--get", "remote.origin.url") or "",
-    }
 
 
 def _is_sqlite_locked_error(exc: BaseException) -> bool:
